bioweb/requirements.txt
asgiref==3.5.1
Django==3.0.3
djangorestframework==3.11.1
factory-boy==3.0.1
Faker==13.11.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
